Bug fix in Updater Service, Change int to long type of timestamp.

+++ src/main/groovy/com/chiorichan/updater/BuildArtifact.java
@@ -435,7 +435,7 @@ public class BuildArtifact
 		{
 			public List<String> affectedPaths;
 			public String commitId;
-			public int timestamp;
+			public long timestamp;
 			public Author author;
 			public String comment;
 			public String date;
